requirements.txt
asgiref==3.8.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
factory_boy==3.3.3
Faker==37.3.0
isort==6.0.1
pillow==11.2.1
PyJWT==2.9.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
